fix: imgui dragint

+++ GearsDebug/include/GLGUI.h
@@ -108,7 +108,7 @@ namespace Glib
         static bool InputVector3(std::string_view label, Vector3* value, bool sameLine = false);
         static bool InputVector4(std::string_view label, Vector4* value, bool sameLine = false);
 
-        static bool DragInt(std::string_view label, int* value, int speed = 1, int min = INT_MIN, int max = INT_MAX, bool sameLine = false);
+        static bool DragInt(std::string_view label, int* value, float speed = 1.0f, int min = INT_MIN, int max = INT_MAX, bool sameLine = false);
         static bool DragFloat(std::string_view label, float* value, float speed = 1.0f, float min = -FLT_MAX, float max = FLT_MAX, bool sameLine = false);
         static bool DragVector2(std::string_view label, Vector2* value, float speed = 1.0f, float min = -FLT_MAX, float max = FLT_MAX, bool sameLine = false);
         static bool DragVector3(std::string_view label, Vector3* value, float speed = 1.0f, float min = -FLT_MAX, float max = FLT_MAX, bool sameLine = false);
